requirements.txt
twitter==1.17.1
gnuplotlib==.15
speedtest-cli==1.0.6